inline service file
